requirements.txt
numpy
opencv-python
Pillow
psutil
pyserial
requests
pyserial
colorsys  # (stdlib - included for completeness)